requirements.txt
python-chess

